Altered the properties of the module to better represent the range of injuries the module produces, also meaning that I can get the injuries 'out' of the dataframe more easily.
I have linked in all the DALY weights for the injuries and these are now assigned appropriately.

I need to create a function that determines the facility which to send a person to, based on their injuries. i.e. if it's a cut say, then maybe this can be dealt with at a health outpost, but if the person has a cut and traumatic brain injury they obviously need to be sent somewhere that can deal with both injuries.
I also need to create a more general HSI interaction event, which consumes resources for all the injuries a person sustains.


Former-commit-id: 4ad9ec38a0e76972203e5e6b30aeb91d26a45a48