Split the RTI_MedicalIntervention event into sub events for each treatment, as the consumable requirements for multiple treatments weren't getting called. This also allows us to see which injuries are treated with consumables and which aren't.
Also changed the way permanent disabilities are modelled, can do it without an additional property, but with more categories for the rt)injury categories.

Still need to make a centralised way to ask whether the person dies without medical intervention as different med ints are happening in multiple events, need a way that ties them together.