Changes the structure of the rt_med_int category, we now track the treatments the person has had in the last month, which allows us to inform which DALYs need to be removed from the dataframe
We can now assign and remove DALYs specific to the injury the person has, with dalys from cast treated fractures and laceration stitches. I will now need to change this for surgeries and burns but it is on the way to.

I need to work out a way to put in a delay for when DALY weights are changed as people will have treatments but still experience the injury for some period of time.