Updated the HSI Start anti_epileptics to request drugs from the health system, improved on how the code was written last so it's now more concise.
I have moved the functions defined in the regular epilepsy event to the top of the epilepsy module. They still function exactly the same, but it's a bit more clearly written. I've also added a lot of comments to what the code is actually doing line by line.

I've also found a potential, small issue with the seizure status transition functions where people could transition multiple seizure states line by line. I've added a todo, which I think I can find a solution for but in the meantime, the model is running as expected