Remove all dance to point `python{,3}` to `python3.8`
It's extremely cumbersome (if system `python3` is `python3.8` then adding the
PPA doesn't work because the `apt_pkg` package isn't available for that version,
and it's very complicated to set `python3` -> `python3.8` only for the `gha`
user in a way that's picked up by the GHA runners) and not worth the effort
trying to make it work sensibly.  This means we have to be explicit about
referencing the version of python we want in the GHA workflows, which is an
acceptable price for this (and being explicit isn't a bad idea anyway).