Some minor improvement to the Ansible playbook for GHA runners (#1149)
* Some minor improvement to the Ansible playbook for GHA runners

Main changes:

* Parametrise the python version to be installed
* Install also the `python3.x-venv` package

* Manually update python3 symlink instead of using alternatives

* Set SSHD config `ClientAliveInterval` to `120`

This avoids connections being killed after ~4.5 minutes of inactivity.

* Set default number of GHA runners

* Symlink `python3` to `python3.8` only for the `gha` user

* Allow setting GHA runners labels from command line

Also, always add the hostname as label, so that we can run workflows exclusively
on a single machine.

* Remove all dance to point `python{,3}` to `python3.8`

It's extremely cumbersome (if system `python3` is `python3.8` then adding the
PPA doesn't work because the `apt_pkg` package isn't available for that version,
and it's very complicated to set `python3` -> `python3.8` only for the `gha`
user in a way that's picked up by the GHA runners) and not worth the effort
trying to make it work sensibly.  This means we have to be explicit about
referencing the version of python we want in the GHA workflows, which is an
acceptable price for this (and being explicit isn't a bad idea anyway).