Update the availability of Prostate specific antigen test
to be the same as that of prostate biopsy. This is based on an assumption agreed with Andrew. All assumptions on availability of consumables not found in the OpenLMIS or HHFA datasets are in the `ResourceFile_hhfa_consumables.xlsx` file in Dropbox - `availability_assumptions` tab. This is all described  in the `Description of derivation of stockout rates from OpenLMIS data_Final.docx` file in Dropbox.