Remove erroneously added "age_exact_years" in test_stunting
These fields in the dataframe are not actually present in the
simulation that is being run in the test (since the demography module
is not loaded).
Additionally, the field is spelt incorrectly in two of three of the
usage instances anyway, so wasn't doing anything even if it was being
setup correctly.