[batch] Use `earlyoom` to terminate a process using too much memory
Linux's built-in system to manage out-of-memory (OOM) errors is unreliable and
non-informative, even if it kicks in (and if often it does too late, when the
machine is already unresponsive) there's no easy way to tell that a process on a
remote machine was terminated because it was using too much memory.  With
`earlyoom` we bypass Linux's OOM management and have better control over what's
printed to screen in case of error.