Responding to aspects of Issue https://github.com/UCL/TLOmodel/issues/1748, in HSI_CardioMetabolicDisorders_Refill_Medication, there is a redone did_not_run function and new never_ran function. These share logic in that, if an appointment to refill medication does not go ahead due to climate,the person will come off of medication but restart it with a certain probability. If the appointment goes ahead, it is rescheduled with a delay that is set by a climate-delay parameter used in the healthcare system.
Also added a check to ensure that the weightloss appointment cannot be scheduled if the person is already on weightloss treatment.
(Concern - is there a more current check to see if someone is currently on weightloss treatment?)

Also included changes in the healthcare system that are in the PR #1604 for illustration of how it would work. So far, only HSI_CardioMetabolicDisorders_Refill_Medication has the weather-related did_not_run and never_ran.